solved forward denoise
